+++ src/client/Entity.cpp
@@ -582,7 +582,7 @@ void EntityEffect::Process()
 				for( int i = 0; i < _amount; i++, angle += angleDiv )
 				{
 					const CVec spread = CVec( sinf(angle), cosf(angle) ) * _radius;
-					const CVec addVel = CVec( sinf(angle - PI/1.5), cosf(angle - PI/1.5) ) * _radius * _speed * _delay * 2.0f;
+					const CVec addVel = CVec( sinf(angle - (float)PI/1.5f), cosf(angle - (float)PI/1.5f) ) * _radius * _speed * _delay * 2.0f;
 					SpawnEntity(ENT_SPARKLE, _fade, pos + spread, vel + addVel, c, NULL);
 				}
 			}
